+++ src/5_banner.js
@@ -116,21 +116,19 @@ banner = function(branch, options, linkData, storage) {
 	linkData['channel'] = linkData['channel'] || 'app banner';
 
 	var doc = options.iframe ? element.contentWindow.document : document;
+
 	if (utils.mobileUserAgent()) {
 		var referringLink = branch._referringLink();
-		if (referringLink && !options['make_new_link']) {
+		if (referringLink && !options.make_new_link) {
 			doc.getElementById('branch-mobile-action').href = referringLink;
 		}
 		else {
-			branch['link'](linkData, function(err, url) {
-				if (err) {
-					// Todo: figure out something good to do here. Maybe a
-					// long link? Or why not always a long link?
-				}
-				else {
-					doc.getElementById('branch-mobile-action').href = url;
-				}
-			});
+			options['open_app'] = options.open_app;
+			options['make_new_link'] = options.make_new_link;
+			branch['deepview'](linkData, options);
+			doc.getElementById('branch-mobile-action').onclick = function() {
+				branch['deepviewCta']();
+			};
 		}
 	}
 	else {