+++ test/1_utils.js
@@ -236,4 +236,36 @@ describe('utils', function() {
 			}
 		});
 	});
+
+	describe('extractDeeplinkPath', function() {
+		it('should return deeplink path for an https:// url', function() {
+			if (testUtils.go('#test:extractDeeplinkPath')) {
+				assert.strictEqual(
+					'abc/def/',
+					utils.extractDeeplinkPath('https://domain.name/abc/def/'),
+					'should extract deeplink path'
+				);
+			}
+		});
+
+		it('should return deeplink path for a url with implicit protocol', function() {
+			if (testUtils.go('#test:extractDeeplinkPath')) {
+				assert.strictEqual(
+					'abc/def/',
+					utils.extractDeeplinkPath('domain.name/abc/def/'),
+					'should extract deeplink path'
+				);
+			}
+		});
+
+		it('should return empty string if there is no deeplink path', function() {
+			if (testUtils.go('#test:extractDeeplinkPath')) {
+				assert.strictEqual(
+					'',
+					utils.extractDeeplinkPath('https://domain.name'),
+					'should extract deeplink path as empty string'
+				);
+			}
+		});
+	});
 });